+++ apps/fortuna/src/chain.rs
@@ -1,2 +1,3 @@
 pub(crate) mod ethereum;
 pub(crate) mod reader;
+pub(crate) mod utils;
