+++ apps/fortuna/src/chain/utils.rs
@@ -1,0 +1,69 @@
+use {
+    crate::chain::ethereum::{InstrumentedSignablePythContract, PythContractCall},
+    anyhow::{anyhow, Result},
+    ethers::middleware::Middleware,
+    std::sync::Arc,
+    tracing,
+};
+
+pub async fn send_and_confirm(contract_call: PythContractCall) -> Result<()> {
+    let call_name = contract_call.function.name.as_str();
+    let pending_tx = contract_call
+        .send()
+        .await
+        .map_err(|e| anyhow!("Error submitting transaction({}) {:?}", call_name, e))?;
+
+    let tx_result = pending_tx
+        .await
+        .map_err(|e| {
+            anyhow!(
+                "Error waiting for transaction({}) receipt: {:?}",
+                call_name,
+                e
+            )
+        })?
+        .ok_or_else(|| {
+            anyhow!(
+                "Can't verify the transaction({}), probably dropped from mempool",
+                call_name
+            )
+        })?;
+
+    tracing::info!(
+        transaction_hash = &tx_result.transaction_hash.to_string(),
+        "Confirmed transaction({}). Receipt: {:?}",
+        call_name,
+        tx_result,
+    );
+    Ok(())
+}
+
+/// Estimate the cost (in wei) of a transaction consuming gas_used gas.
+pub async fn estimate_tx_cost(
+    contract: Arc<InstrumentedSignablePythContract>,
+    use_legacy_tx: bool,
+    gas_used: u128,
+) -> Result<u128> {
+    let middleware = contract.client();
+
+    let gas_price: u128 = if use_legacy_tx {
+        middleware
+            .get_gas_price()
+            .await
+            .map_err(|e| anyhow!("Failed to fetch gas price. error: {:?}", e))?
+            .try_into()
+            .map_err(|e| anyhow!("gas price doesn't fit into 128 bits. error: {:?}", e))?
+    } else {
+        // This is not obvious but the implementation of estimate_eip1559_fees in ethers.rs
+        // for a middleware that has a GasOracleMiddleware inside is to ignore the passed-in callback
+        // and use whatever the gas oracle returns.
+        let (max_fee_per_gas, max_priority_fee_per_gas) =
+            middleware.estimate_eip1559_fees(None).await?;
+
+        (max_fee_per_gas + max_priority_fee_per_gas)
+            .try_into()
+            .map_err(|e| anyhow!("gas price doesn't fit into 128 bits. error: {:?}", e))?
+    };
+
+    Ok(gas_price * gas_used)
+}
